Refactor provider package tooling from global constants to provider dependencies
