+++ dev/breeze/src/airflow_breeze/global_constants.py
@@ -20,25 +20,16 @@ Global constants that are used by all other Breeze components.
 
 from __future__ import annotations
 
-import json
 import platform
-import subprocess
-from collections.abc import Generator
 from enum import Enum
-from functools import cache
-from pathlib import Path
-from threading import Lock
 
 from airflow_breeze.utils.functools_cache import clearable_cache
 from airflow_breeze.utils.host_info_utils import Architecture
 from airflow_breeze.utils.path_utils import (
     AIRFLOW_CORE_SOURCES_PATH,
     AIRFLOW_CTL_SOURCES_PATH,
-    AIRFLOW_PYPROJECT_TOML_FILE_PATH,
     AIRFLOW_ROOT_PATH,
     AIRFLOW_TASK_SDK_SOURCES_PATH,
-    PROVIDER_DEPENDENCIES_JSON_HASH_PATH,
-    PROVIDER_DEPENDENCIES_JSON_PATH,
 )
 
 PUBLIC_AMD_RUNNERS = '["ubuntu-22.04"]'
@@ -660,80 +651,9 @@ PROVIDER_RUNTIME_DATA_SCHEMA_PATH = AIRFLOW_CORE_SOURCES_PATH / "airflow" / "pro
 ALL_PYPROJECT_TOML_FILES = []
 
 
-def get_all_provider_pyproject_toml_provider_yaml_files() -> Generator[Path, None, None]:
-    pyproject_toml_content = AIRFLOW_PYPROJECT_TOML_FILE_PATH.read_text().splitlines()
-    in_workspace = False
-    for line in pyproject_toml_content:
-        trimmed_line = line.strip()
-        if not in_workspace and trimmed_line.startswith("[tool.uv.workspace]"):
-            in_workspace = True
-        elif in_workspace:
-            if trimmed_line.startswith("#"):
-                continue
-            if trimmed_line.startswith('"'):
-                path = trimmed_line.split('"')[1]
-                ALL_PYPROJECT_TOML_FILES.append(AIRFLOW_ROOT_PATH / path / "pyproject.toml")
-                if trimmed_line.startswith('"providers/'):
-                    yield AIRFLOW_ROOT_PATH / path / "pyproject.toml"
-                    yield AIRFLOW_ROOT_PATH / path / "provider.yaml"
-            elif trimmed_line.startswith("]"):
-                break
-
-
-_regenerate_provider_deps_lock = Lock()
-
 UPDATE_PROVIDER_DEPENDENCIES_SCRIPT = (
     AIRFLOW_ROOT_PATH / "scripts" / "ci" / "prek" / "update_providers_dependencies.py"
 )
-
-
-@cache  # Note: using functools.cache to avoid multiple dumps in the same run
-def regenerate_provider_dependencies_once() -> None:
-    """Run provider dependencies regeneration once per interpreter execution.
-
-    This function is safe to call multiple times from different modules; the
-    underlying command will only run once. If the underlying command fails the
-    CalledProcessError is propagated to the caller.
-    """
-    with _regenerate_provider_deps_lock:
-        # Run the regeneration command from the repository root to ensure correct
-        # relative paths if the script expects to be run from AIRFLOW_ROOT_PATH.
-        subprocess.check_call(
-            ["uv", "run", UPDATE_PROVIDER_DEPENDENCIES_SCRIPT.as_posix()], cwd=AIRFLOW_ROOT_PATH
-        )
-
-
-def _calculate_provider_deps_hash():
-    import hashlib
-
-    hasher = hashlib.sha256()
-    for file in sorted(get_all_provider_pyproject_toml_provider_yaml_files()):
-        hasher.update(file.read_bytes())
-    return hasher.hexdigest()
-
-
-@cache
-def get_provider_dependencies() -> dict:
-    if not PROVIDER_DEPENDENCIES_JSON_PATH.exists():
-        calculated_hash = _calculate_provider_deps_hash()
-        PROVIDER_DEPENDENCIES_JSON_HASH_PATH.write_text(calculated_hash)
-        # We use regular print there as rich console might not be initialized yet here
-        print("Regenerating provider dependencies file")
-        regenerate_provider_dependencies_once()
-    return json.loads(PROVIDER_DEPENDENCIES_JSON_PATH.read_text())
-
-
-def generate_provider_dependencies_if_needed():
-    if not PROVIDER_DEPENDENCIES_JSON_PATH.exists() or not PROVIDER_DEPENDENCIES_JSON_HASH_PATH.exists():
-        get_provider_dependencies.cache_clear()
-        get_provider_dependencies()
-    else:
-        calculated_hash = _calculate_provider_deps_hash()
-        if calculated_hash.strip() != PROVIDER_DEPENDENCIES_JSON_HASH_PATH.read_text().strip():
-            # Force re-generation
-            PROVIDER_DEPENDENCIES_JSON_PATH.unlink(missing_ok=True)
-            get_provider_dependencies.cache_clear()
-            get_provider_dependencies()
 
 
 DEVEL_DEPS_PATH = AIRFLOW_ROOT_PATH / "generated" / "devel_deps.txt"
